little formatting

+++ src/BitBases.cpp
@@ -110,7 +110,7 @@ namespace BitBases {
         // bit 15-17: white pawn R_7 - rank (from R_7 to R_2)
         inline u32 index (Color c, Square bk_sq, Square wk_sq, Square wp_sq)
         {
-            return wk_sq | (bk_sq << 6) | (c << 12) | (_file (wp_sq) << 13) | ((i32(R_7) - i32(_rank (wp_sq))) << 15);
+            return wk_sq | (bk_sq << 6) | (c << 12) | (_file (wp_sq) << 13) | ((R_7 - _rank (wp_sq)) << 15);
         }
 
         template<Color Own>
